// src/pages/explore/explore.types.ts
export interface Category {
  id: string
  name: string
}

export interface Rating {
  id: string
  rate: number
  description: string
  created_at: Date
  book_id: string
  user_id: string
}

export interface Book {
  id: string
  name: string
  author: string
  summary: string
  cover_url: string
  total_pages: number
  created_at: Date
  ratings: Rating[]
  categories: { category: Category }[]
}